requirements.txt
Django==4.1.6
whitenoise==6.3.0
pandas~=1.5.3
djangorestframework~=3.14.0
scikit-learn==1.2.1